requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
htbuilder>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
pymupdf>=1.23.0
langfuse>=2.0.0
